Export type definitions for button and item props
If wrapping the library component for standardization of menu usage
patterns within an app, it can be useful to keep the wrapper thin. A
good way to do that is to re-export parts of the props of the
underlying component. However, these types are not exported right now.

Export the type definitions for button and item props for the menu.